Refactor: add App.js
